+++ examples/with-svelte/client/src/main.ts
@@ -1,0 +1,12 @@
+/** @format */
+
+import App from "./App.svelte";
+
+const app = new App({
+    target: document.body,
+    props: {
+        name: "from supertokens",
+    },
+});
+
+export default app;